Refactor: Role 디렉토리 이동
