--- src/main/kotlin/com/quizit/authentication/domain/Role.kt
@@ -1,6 +1,0 @@
-package com.quizit.authentication.domain
-
-enum class Role {
-    ADMIN,
-    USER
-}
