+++ src/main/kotlin/com/quizit/authentication/domain/enum/Role.kt
@@ -1,0 +1,6 @@
+package com.quizit.authentication.domain.enum
+
+enum class Role {
+    ADMIN,
+    USER
+}
